# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted note-taking application API. Health probes are available at root paths `/livez` and `/readyz` (outside `/api/v1`).",
        "title": "Jot API",
        "contact": {
            "name": "Jot Project",
            "url": "https://github.com/hanzei/jot"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/labels": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labels"
                ],
                "summary": "List all labels for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LabelListResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labels"
                ],
                "summary": "Create a label, or return the existing one when no ID is supplied",
                "parameters": [
                    {
                        "description": "Label name and optional client-supplied ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddLabelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "an existing label with that name was returned unchanged",
                        "schema": {
                            "$ref": "#/definitions/models.Label"
                        }
                    },
                    "201": {
                        "description": "a new label was created",
                        "schema": {
                            "$ref": "#/definitions/models.Label"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "label already exists",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "label name too long (label_name_too_long)",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
        "/labels/counts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labels"
                ],
                "summary": "Get note counts per label for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LabelCountsResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
        "/labels/{id}": {
            "delete": {
                "tags": [
                    "labels"
                ],
                "summary": "Delete a label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Label ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "label not found",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labels"
                ],
                "summary": "Rename a label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Label ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New label name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RenameLabelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Label"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "label not found",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "label name too long (label_name_too_long)",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
...
        "/notes/{id}/labels": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labels"
                ],
                "summary": "Add a label to a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Label name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddLabelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Note"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "no access to note",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "label not found",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "label name too long (label_name_too_long)",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
        "/notes/{id}/labels/{label_id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labels"
                ],
                "summary": "Remove a label from a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Label ID",
                        "name": "label_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Note"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "no access to note",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "apierr.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is stable and machine-readable; clients switch on it.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/apierr.Code"
                        }
                    ],
                    "example": "not_found"
                },
                "message": {
                    "description": "Message is human-readable English, for logs and as a fallback. Masked\nto \"internal server error\" for 5xx responses.",
                    "type": "string",
                    "example": "note not found"
                }
            }
        },
        "apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/apierr.ErrorDetail"
                }
            }
        },
        "handlers.AddLabelRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID is an optional client-supplied label ID. When provided it is used as the\nlabel's primary key so an offline-created label can be replayed idempotently:\na replay whose original create already committed is rejected with 409\ninstead of inserting a duplicate. When empty the server generates one.",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LabelCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "label_id": {
                    "type": "string"
                }
            }
        },
        "handlers.LabelCountsResponse": {
            "type": "object",
            "properties": {
                "counts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.LabelCount"
                    }
                }
            }
        },
        "handlers.LabelListResponse": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Label"
                    }
                }
            }
        },
...
        "handlers.RenameLabelRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Label": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Note": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "checked_items_collapsed": {
                    "type": "boolean"
                },
                "color": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NoteImage"
                    }
                },
                "is_shared": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NoteItem"
                    }
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Label"
                    }
                },
                "note_type": {
                    "$ref": "#/definitions/models.NoteType"
                },
                "pinned": {
                    "type": "boolean"
                },
                "position": {
                    "type": "integer"
                },
                "shared_with": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NoteShare"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "version": {
                    "description": "Version is an optimistic-concurrency counter bumped on every shared-content\n(title/content) change. Clients echo the version their edit was based on as\nbase_version on update so a stale write can be rejected (issue #489).",
                    "type": "integer"
                }
            }
        },
        "models.NoteImage": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "jot_session",
            "in": "header"
        }
    }
}